requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.34.0

# Google ADK and Vertex AI
google-adk==0.3.0
google-cloud-aiplatform==1.87.0
google-auth==2.35.0
google-genai==1.11.0

# Qdrant
qdrant-client==1.12.0

# HTTP client
httpx==0.28.1

# Database
libsql-client==0.3.0

# Scheduler
apscheduler==3.10.4

# Search
google-search-results==2.4.2

# Utils
python-dotenv==1.0.1
pydantic==2.9.2

# Logging
loguru==0.7.2
